pyproject.toml
[tool.poetry]
name = "api-test"
version = "0.1.0"
description = "Test task to implement CRUD with API"
authors = ["Anton <antonukr@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
